Detect outlier loci based on their contribution to a RDA model

The rdadapt function performs redundancy analysis and computes p-values to test for outliers based on loci extremeness along a distribution of Mahalanobis distances estimated between each locus and the center of the RDA space using a certain number of axes (K). rdadapt accommodates individual genotypes or allele frequencies.

rdadapt(RDA, K, scores)

# S4 method for rda,ANY,ANY
rdadapt(RDA, K)

# S4 method for missing,missing,data.frame
rdadapt(scores)

# S4 method for missing,missing,matrix
rdadapt(scores)

Arguments

RDA

a RDA model from which to extract loci loadings

K

an integer specifying the number of RDA axes to use for the detection

Value

A data.frame containing :

p.values : the p-value associated with each locus

q.values : the q-value associated with each locus, estimated using the q.values package and allowing to use a FDR (False Discovery Rate) approach instead of a p-value threshold to identify outliers.

Details

First, Mahalanobis distances are computed for all genetic marker using a robust estimate of both mean and covariance matrix between the K RDA vectors of loadings. Then, to compute p-values, Mahalanobis distances are divided by a genomic inflation factor (gif), giving a scaled statistic that should follow a chi-squared distribution with K degrees of freedom.

Author

Thibaut Capblancq